Update jquery.lazyloadxt.bg.js
If background plugin is used at the same time as DOM elements that is set to lazy load, the background plugin incorrectly also injects a background-image style to each element. This causes a lot of undefined image loading

```style="background-image: url(http://magento.zabadaclean.biz/undefined);"```

This fix ads a check to determine that an element has a valid defined bgAttr set.